Moved InventoryUtils.cs
