+++ stardew-access/Patches/DonationMenuPatches/FieldOfficeMenuPatch.cs
@@ -1,4 +1,5 @@
 using StardewValley;
+using stardew_access.Features;
 using StardewValley.Menus;
 
 namespace stardew_access.Patches
