+++ stardew-access/Patches/DonationMenuPatches/MuseumMenuPatch.cs
@@ -1,5 +1,6 @@
 using Microsoft.Xna.Framework;
 using StardewValley;
+using stardew_access.Features;
 using StardewValley.Locations;
 using StardewValley.Menus;
