+++ stardew-access/Patches/GameMenuPatches/CraftingPagePatch.cs
@@ -1,4 +1,5 @@
 using StardewValley;
+using stardew_access.Features;
 using StardewValley.Menus;
 using StardewValley.Objects;
