+++ stardew-access/Patches/IClickableMenuPatch.cs
@@ -1,3 +1,4 @@
+using stardew_access.Features;
 using StardewValley;
 using StardewValley.Menus;
